use float32, add test

+++ napari/_vispy/utils/gl.py
@@ -90,7 +90,7 @@ def fix_data_dtype(data):
         return data
     else:
         try:
-            dtype = dict(i=np.float64, f=np.float64, u=np.uint16, b=np.uint8)[
+            dtype = dict(i=np.float32, f=np.float32, u=np.uint16, b=np.uint8)[
                 dtype.kind
             ]
         except KeyError:  # not an int or float
